+++ src/temp.h
@@ -6,7 +6,7 @@ int io_uring_setup(unsigned entries, io_uring_params *params);
 int io_uring_enter(int ring_fd, unsigned to_submit, unsigned min_complete, unsigned flags);
 int io_uring_register(unsigned ring_fd, unsigned op, void *arg, unsigned nr_args);
 
-int setup_server_socket(std::string address, int port);
+int setup_server_socket(std::string address, unsigned short port);
 bool tune_socket(int fd, int buffer_size = 4 * 1024 * 1024, bool quick_ack = true, bool no_delay = true);
 
 unsigned long pack_fd_index_opcode(int fd, unsigned index, unsigned char opcode);
